+++ packages/yume-engine/src/bundle/index.ts
@@ -1,0 +1,1 @@
+export * from './xml-plist-parser.js';